Dynamically set metadata from store value
