+++ package.json
@@ -35,6 +35,7 @@
     "remark-rehype": "^10.0.0",
     "svelte": "^3.46.4",
     "svelte-check": "^2.4.5",
+    "svelte-meta-tags": "^2.4.0",
     "svelte-preprocess": "^4.10.4",
     "to-vfile": "^7.2.0",
     "tslib": "^2.3.1",
