+++ src/routes/blog/[slug].ts
@@ -1,18 +1,5 @@
 import { loadPost } from './_posts';
 
-export interface BlogPost {
-	metadata: {
-		title: string;
-		date: string;
-		draft: boolean;
-		description: string;
-		tags: string[];
-	};
-	html: string;
-	fileName: string;
-	slug: string;
-}
-
 /** @type {import('./[slug]').RequestHandler} */
 export async function get({ params }) {
 	// the `slug` parameter is available because
